requirements.txt
autopep8==1.5.2
click==7.1.2
fakeredis==1.4.1
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==4.0.2
Pillow==7.1.2
pycodestyle==2.6.0
redis==3.5.2
rq==1.4.1
six==1.14.0
sortedcontainers==2.1.0
Werkzeug==1.0.1
